fixed teardown bug in some startup failure

+++ examples/docker/dockerized.ts
@@ -27,11 +27,15 @@ type ContainerMetadata = {
 
 function createDockerizedService(opts: DockerOptions): Service {
   const name = opts.name || uuid();
+  let started = false;
 
   return {
     id: name,
     start: async (): Promise<ContainerMetadata> => {
+      logger.debug('starting container %s...', name);
       await executeCommand(interpret(name, opts));
+      started = true;
+      logger.debug('container %s started', name);
 
       return {
         id: name,
@@ -43,7 +47,11 @@ function createDockerizedService(opts: DockerOptions): Service {
       };
     },
     stop: async () => {
-      await executeCommand(`docker kill ${name}`);
+      if (started) {
+        logger.debug('killing container %s...', name);
+        await executeCommand(`docker kill ${name}`);
+        logger.debug('container %s killed', name);
+      }
     }
   };
 }
@@ -84,14 +92,16 @@ async function executeCommand(cmd: string): Promise<number> {
   logger.debug('Running: %s', cmd);
   return new Promise<number>((resolve, reject) => {
     const p = child_process.spawn(cmd, { stdio: 'inherit', shell: true });
+    const onExit = (code: number) => {
+      if (code !== 0) {
+        reject(new Error(`Command exit code: ${code}`));
+      }
+      resolve(code);
+    };
 
     p.on('error', reject);
-    p.on('exit', code => {
-      resolve(code);
-    });
-    p.on('close', code => {
-      resolve(code);
-    });
+    p.on('exit', onExit);
+    p.on('close', onExit);
   });
 }
